Development

+++ server.js
@@ -123,6 +123,7 @@ const server = http.createServer((req, res) => {
 			const collection = postMessage.collection;
 			const isExternalNetwork = postMessage.isExternalNetwork;
 			if(!src || !prompt || !collection) return "failed";
+			const d = new Date();
 
 			const urlArr = src.split("/");
 
@@ -134,6 +135,12 @@ const server = http.createServer((req, res) => {
 			const safeFileName = utils.makeFileNameSafe(prompt);
 			const newFileName = `${safeFileName}.${fileExtension}`;
 			savePath += newFileName;
+
+			if(!fs.existsSync(src)) {
+				console.log(d.toLocaleTimeString() + `-> failed to save preview: file "${src}" not found`);
+
+				return "failed";
+			}
 
 			fs.copyFile(src, savePath, (err) => {
 				if (err) throw err;
@@ -185,7 +192,6 @@ const server = http.createServer((req, res) => {
 				}
 			}
 
-			const d = new Date();
 			console.log(d.toLocaleTimeString() + "-> updated preview and data for: " + collection);
 			
 			return "ok";
@@ -451,6 +457,7 @@ const server = http.createServer((req, res) => {
 			const style = postMessage.style;
 			const collection = postMessage.collection;
 			if(!src || !style || !collection) return "failed";
+			const d = new Date();
 
 			const urlArr = src.split("/");
 
@@ -467,6 +474,12 @@ const server = http.createServer((req, res) => {
 			const newFileName = `${safeFileName}.${fileExtension}`;
 			savePath += newFileName;
 
+			if(!fs.existsSync(src)) {
+				console.log(d.toLocaleTimeString() + `-> failed to save preview: file "${src}" not found`);
+
+				return "failed";
+			}
+
 			fs.copyFile(src, savePath, (err) => {
 				if (err) throw err;
 			});
@@ -481,9 +494,88 @@ const server = http.createServer((req, res) => {
 				fs.writeFileSync(pathToDataFile, JSON.stringify(stylesJSON, null, "\t"));
 			}
 
-			const d = new Date();
 			console.log(d.toLocaleTimeString() + "-> updated style preview for: " + style);
 			
+			return "ok";
+		});
+	}
+
+	if(req.method === "POST" && req.url === "/newCollection") {
+		let body = "";
+
+		req.on("data", chunk => {
+			body += chunk.toString();
+		});
+
+		req.on("end", () => {
+			const postMessage = JSON.parse(body);
+			const id = postMessage.id;
+			const mode = postMessage.mode;
+			if(!id || !mode) return "failed";
+			const d = new Date();
+
+			const pathToCollection = promptsCataloguePath + path.sep + id + path.sep;
+
+			if(fs.existsSync(pathToCollection)) {
+				console.log(d.toLocaleTimeString() + `-> failed to create new collection: id ${id} already exists.`);
+
+				return "false";
+			}
+
+			fs.mkdirSync(pathToCollection);
+			fs.mkdirSync(pathToCollection + path.sep + "preview");
+
+			const metaJSON = {
+				format: mode
+			};
+
+			fs.writeFileSync(pathToCollection + path.sep + "meta.json", JSON.stringify(metaJSON, null, "\t"));
+
+			if(mode === "expanded") {
+				fs.mkdirSync(pathToCollection + path.sep + "prompts");
+				const orderJSON = [];
+				fs.writeFileSync(pathToCollection + path.sep + "order.json", JSON.stringify(orderJSON, null, "\t"));
+
+			} else {
+				const dataJSON = [];
+				fs.writeFileSync(pathToCollection + path.sep + "data.json", JSON.stringify(dataJSON, null, "\t"));
+			}
+
+			console.log(d.toLocaleTimeString() + `-> created new prompts collection: ${id}.`);
+
+			return "ok";
+		});
+	}
+
+	if(req.method === "POST" && req.url === "/newStylesCollection") {
+		let body = "";
+
+		req.on("data", chunk => {
+			body += chunk.toString();
+		});
+
+		req.on("end", () => {
+			const postMessage = JSON.parse(body);
+			const id = postMessage.id;
+			if(!id) return "failed";
+			const d = new Date();
+
+			const pathToCollection = stylesCataloguePath + path.sep + id + path.sep;
+
+			if(fs.existsSync(pathToCollection)) {
+				console.log(d.toLocaleTimeString() + `-> failed to create new styles collection: id ${id} already exists.`);
+
+				return "false";
+			}
+
+			fs.mkdirSync(pathToCollection);
+			fs.mkdirSync(pathToCollection + path.sep + "preview");
+
+			const dataJSON = [];
+			fs.writeFileSync(pathToCollection + path.sep + "data.json", JSON.stringify(dataJSON, null, "\t"));
+
+			console.log(d.toLocaleTimeString() + `-> created new styles collection: ${id}.`);
+
 			return "ok";
 		});
 	}
